Rewriting Point Code
Rewrote some code that determines where each Point that an entity might use to look at, move to, etc. Next plan for update is to add in a general controller script for the player and enemies so that they can have a central "mind" to control them and decide what they do. This will act similarly to a motherboard in how it directs each input and output to the different parts of the computer.