Modularity for General Entity Scripts
Created super classes for scripts related to movement, attacks, and bullet types. Begun setup for current scripts to use this format so that the main entity controller scripts can easily trade out the current scripts the entity is using to easily switch between the finite states the AI is comprised of.